Fixed only bugs that are revealed when doing actual joins
While the initial commits did produce dataframes and tables that looked sane, only when we do joins can we tell what's what.  This check-in fixed some major bugs and oversights, while also bringing in the latest idioms from the trino ingenstion notebooks from Red Hat.

Another major update was producing 10 years worth of GLEIF dictionaries (because naming conventions drifted so much that a dictionary matching 5000+ names at one end of the decade could barely match 100 on the other--and vice versa!).

Unfinished business remains: the handling of "confidential" numbers if spreadsheets where floats are supposed to be is one biggie.